requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
pillow==11.1.0
piexif==1.1.3
parse==1.20.2
PyYAML==6.0.2
handymatt>=1.1.0
fastapi==0.115.13
uvicorn==0.34.3
